+++ src/tech-stack/application/dtos/response.dto.ts
@@ -1,3 +1,3 @@
-import type { TechStackCategory } from "@/tech-stack";
+import type { TechStackCategory } from "@/tech-stack/application/types";
 
 export type FetchTechStackResponseDto = TechStackCategory[];
