+++ src/tech-stack/index.ts
@@ -1,21 +1,7 @@
-import type { VoyageMember } from "@/features/application/types";
-
-export interface TechStackCategory {
-  id: number;
-  name: string;
-  description: string;
-  teamTechStackItems: TechStackItem[];
-}
-
-export interface TechStackItem {
-  id: number;
-  name: string;
-  isSelected: boolean;
-  teamTechStackItemVotes: TechStackItemVotes[];
-}
-
-export interface TechStackItemVotes {
-  votedBy: {
-    member: VoyageMember;
-  };
-}
+export { TechStackClientAdapter } from "./adapters/primary/tech-stack-client-adapter";
+export { FetchTechStackClientRequestDto } from "./application/dtos/request.dto";
+export {
+  TechStackCategory,
+  TechStackItem,
+  TechStackItemVotes,
+} from "./application/types/index";
